+++ src/helpers/arrayHelper.ts
@@ -4,12 +4,12 @@
 import { ObjectHelper } from "./objectHelper";
 
 export class ArrayHelper {
-    public static addRemove(arr: any[], key: any, add: boolean): void {
-        if (arr !== undefined && arr !== null && key !== undefined && key !== null) {
-            const idx = arr.indexOf(key);
+    public static addRemove<T>(arr: T[], object: T, add: boolean, matcher?: (object: T, item: T) => boolean): void {
+        if (arr !== undefined && arr !== null && object !== undefined && object !== null) {
+            const idx = matcher ? arr.findIndex(item => matcher(object, item)) : arr.indexOf(object);
             if (add) {
                 if (idx < 0) {
-                    arr.push(key);
+                    arr.push(object);
                 }
             } else {
                 if (idx >= 0) {